Aprobado

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: emedina- <emedina-@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/28 09:38:35 by emedina-          #+#    #+#             */
-/*   Updated: 2023/05/16 12:36:01 by emedina-         ###   ########.fr       */
+/*   Updated: 2023/05/17 19:38:53 by emedina-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	size_t i;
-	const unsigned char *str1;
-	const unsigned char *str2;
+	size_t				i;
+	const unsigned char	*str1;
+	const unsigned char	*str2;
 
 	str1 = s1;
 	str2 = s2;
@@ -25,7 +25,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	{
 		if (str1[i] != str2[i])
 		{
-			return (size_t)(str1[i] - str2[i]);
+			return ((unsigned int)(str1[i] - str2[i]));
 		}
 		i++;
 	}
